Former-commit-id: ae0408b2a6e4377c681a6ad993d7281064d3ebae

+++ src/source_controller/coreservice/core/process/process_instance_relation.go
@@ -16,6 +16,7 @@ import (
 	"configcenter/src/common"
 	"configcenter/src/common/blog"
 	"configcenter/src/common/errors"
+	"configcenter/src/common/eventclient"
 	"configcenter/src/common/metadata"
 	"configcenter/src/source_controller/coreservice/core"
 )
@@ -43,19 +44,26 @@ func (p *processOperation) CreateProcessInstanceRelation(ctx core.ContextParams,
 		blog.Errorf("CreateProcessInstanceRelation failed, service instance id invalid, code: %d, err: %+v, rid: %s", common.CCErrCommParamsInvalid, err, ctx.ReqID)
 		return nil, ctx.Error.CCErrorf(common.CCErrCommParamsInvalid, "service_instance_id")
 	}
-
-	// validate service category id field
-	_, err = p.GetServiceInstance(ctx, relation.ServiceInstanceID)
-	if err != nil {
-		blog.Errorf("CreateProcessInstanceRelation failed, service instance id invalid, code: %d, err: %+v, rid: %s", common.CCErrCommParamsInvalid, err, ctx.ReqID)
-		return nil, ctx.Error.CCErrorf(common.CCErrCommParamsInvalid, "service_instance_id")
-	}
 	// TODO: asset bizID == category.BizID
 
 	relation.SupplierAccount = ctx.SupplierAccount
 	if err := p.dbProxy.Table(common.BKTableNameProcessInstanceRelation).Insert(ctx.Context, &relation); nil != err {
 		blog.Errorf("CreateProcessInstanceRelation failed, mongodb failed, table: %s, relation: %+v, err: %+v, rid: %s", common.BKTableNameProcessInstanceRelation, relation, err, ctx.ReqID)
 		return nil, ctx.Error.CCErrorf(common.CCErrCommDBInsertFailed)
+	}
+
+	// push process instance relation create event
+	event := eventclient.NewEventWithHeader(ctx.Header)
+	event.EventType = metadata.EventTypeRelation
+	event.ObjType = metadata.EventObjTypeProcModule
+	event.Action = metadata.EventActionCreate
+	event.Data = []metadata.EventData{
+		{CurData: relation},
+	}
+	err = p.eventCli.Push(ctx, event)
+	if err != nil {
+		blog.Errorf("process instance relation event push failed, error:%v, rid: %s", err, ctx.ReqID)
+		return nil, ctx.Error.CCErrorf(common.CCErrCoreServiceEventPushEventFailed)
 	}
 	return relation, nil
 }
@@ -92,11 +100,35 @@ func (p *processOperation) UpdateProcessInstanceRelation(ctx core.ContextParams,
 		return nil, err
 	}
 
-	// do update
 	filter := map[string]int64{"process_id": processInstanceID}
+
+	preData := make([]map[string]interface{}, 0)
+	if err := p.dbProxy.Table(common.BKTableNameProcessInstanceRelation).Find(filter).All(ctx.Context, &preData); nil != err {
+		blog.Errorf("UpdateProcessInstanceRelation failed, find relation in mongodb failed, table: %s, filter: %+v, err: %+v, rid: %s", common.BKTableNameProcessInstanceRelation, filter, err, ctx.ReqID)
+		return nil, ctx.Error.CCErrorf(common.CCErrCommDBSelectFailed)
+	}
+
+	// do update
 	if err := p.dbProxy.Table(common.BKTableNameProcessInstanceRelation).Update(ctx, filter, relation); nil != err {
 		blog.Errorf("UpdateProcessInstanceRelation failed, mongodb failed, table: %s, filter: %+v, relation: %+v, err: %+v, rid: %s", common.BKTableNameProcessInstanceRelation, filter, relation, err, ctx.ReqID)
 		return nil, ctx.Error.CCErrorf(common.CCErrCommDBUpdateFailed)
+	}
+
+	// push process instance relation update event
+	var eventArr []*metadata.EventInst
+	for _, data := range preData {
+		event := eventclient.NewEventWithHeader(ctx.Header)
+		event.EventType = metadata.EventTypeRelation
+		event.ObjType = metadata.EventObjTypeProcModule
+		event.Action = metadata.EventActionUpdate
+		event.Data = []metadata.EventData{
+			{PreData: data, CurData: relation},
+		}
+		eventArr = append(eventArr, event)
+	}
+	if err := p.eventCli.Push(ctx, eventArr...); err != nil {
+		blog.Errorf("process instance relation event push failed, error:%v, rid: %s", err, ctx.ReqID)
+		return nil, ctx.Error.CCErrorf(common.CCErrCoreServiceEventPushEventFailed)
 	}
 	return relation, nil
 }
@@ -185,9 +217,33 @@ func (p *processOperation) DeleteProcessInstanceRelation(ctx core.ContextParams,
 		return ctx.Error.CCErrorf(common.CCErrCommParametersCountNotEnough)
 	}
 
+	preData := make([]map[string]interface{}, 0)
+	if err := p.dbProxy.Table(common.BKTableNameProcessInstanceRelation).Find(deleteFilter).All(ctx.Context, &preData); nil != err {
+		blog.Errorf("DeleteProcessInstanceRelation failed, find relation in mongodb failed, table: %s, filter: %+v, err: %+v, rid: %s", common.BKTableNameProcessInstanceRelation, deleteFilter, err, ctx.ReqID)
+		return ctx.Error.CCErrorf(common.CCErrCommDBSelectFailed)
+	}
+
 	if err := p.dbProxy.Table(common.BKTableNameProcessInstanceRelation).Delete(ctx, deleteFilter); nil != err {
 		blog.Errorf("DeleteProcessInstanceRelation failed, mongodb failed, table: %s, filter: %+v, err: %+v, rid: %s", common.BKTableNameProcessInstanceRelation, deleteFilter, err, ctx.ReqID)
 		return ctx.Error.CCErrorf(common.CCErrCommDBDeleteFailed)
+	}
+
+	// push process instance relation delete event
+	var eventArr []*metadata.EventInst
+	for _, data := range preData {
+		event := eventclient.NewEventWithHeader(ctx.Header)
+		event.EventType = metadata.EventTypeRelation
+		event.ObjType = metadata.EventObjTypeProcModule
+		event.Action = metadata.EventActionDelete
+		event.Data = []metadata.EventData{
+			{PreData: data},
+		}
+		eventArr = append(eventArr, event)
+	}
+	err := p.eventCli.Push(ctx, eventArr...)
+	if err != nil {
+		blog.Errorf("process instance relation event push failed, error:%v, rid: %s", err, ctx.ReqID)
+		return ctx.Error.CCErrorf(common.CCErrCoreServiceEventPushEventFailed)
 	}
 	return nil
 }
